Add support for V2 devices.
This commit adds support for V2 devices which use a different protocol
to the original streamdeck.

The write paths for all four device models are now unified.

All devices are now adressed zero-indexed starting top-left, for
consistency.

The written images are now flipped correctly for all models, and for
this I updated the `image` dependency as well. The various public button
setting functions will internally convert to the correct format.
`set_button_image` now takes a `DynamicImage` instead of a raw buffer
and applies the relevant transformations to get the image into the
correct device format.

Tested on Original and XL streamdecks, but not the Original V2 or Mini
as I don't have those available for testing.