Map buttons when reading state
When reading button state, we currently get inconsistent ordering and
for the XL we don't get all keys. We also get some extra garbage at the
start of the report.

Here we filter that out and remap the buttons so they match a 0-indexed
left-to-right order on the Streamdeck, for all models.

With this, `read_buttons` now returns a vector of all current button
states. The array has one element for each key, with either a 1 for
pressed or 0 for not pressed.