Add `connect_with_hid` to StreamDeck
Since only one HidApi instance can exist at any given
point and time, having an option to manage the lifecycle
of that instance outside of StreamDeck is useful.

The example I needed this new function for was to enable reconnection
scenarios, as I couldn't call `connect()` while I still had an instance
of StreamDeck in scope, because `HidApi::new()?` would return
an Error.